movie-recommendation-system/requirements.txt
# 项目依赖
django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
requests==2.31.0
python-dotenv==1.0.0
django-environ==0.11.2
django-crispy-forms==2.0
django-filter==23.3
django-debug-toolbar==4.2.0
django-extensions==3.2.3
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
ipython==8.14.0
tqdm==4.65.0
pillow==10.0.0
